refactor: consolidate ashrae_140 into ashrae40 module
- Move high_mass.rs from ashrae_140 to ashrae140
- Add pub mod high_mass to ashrae40/mod.rs
- Remove pub mod ashrae_140 from validation/mod.rs
- Update lib.rs to use validation::ashrae40::high_mass

This resolves the duplicate ashrae directories issue (ARCH-005)
