+++ src/lib.rs
@@ -1600,7 +1600,7 @@ fn fluxion(_py: Python, m: &Bound<'_, PyModule>) -> PyResult<()> {
 }
 
 // Re-export ASHRAE 140 validation models
-pub use validation::ashrae_140::high_mass;
+pub use validation::ashrae140::high_mass;
 
 // Tests for core physics engine (no Python bindings required)
 #[cfg(test)]
