--- src/validation/ashrae_140/mod.rs
@@ -1,6 +1,0 @@
-//! ASHRAE Standard 140 validation test cases.
-//!
-//! This module contains implementations of ASHRAE Standard 140 test cases
-//! for validating building energy simulation accuracy.
-
-pub mod high_mass;
